fastapi/requirements.txt
fastapi
uvicorn
pymongo
transformers
torch
Pillow
env
